fix cli issue
